# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: COMET LKW-Verladungsverwaltung API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      parameters:
        - name: speditionId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete (deactivate) user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deactivated
  /speditionen:
    get:
      operationId: listSpeditionen
      tags: [speditionen]
      summary: List freight companies
      responses:
        "200":
          description: Speditionen list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Spedition"
    post:
      operationId: createSpedition
      tags: [speditionen]
      summary: Create freight company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeditionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Spedition"
  /speditionen/{id}:
    get:
      operationId: getSpedition
      tags: [speditionen]
      summary: Get freight company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Spedition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Spedition"
    patch:
      operationId: updateSpedition
      tags: [speditionen]
      summary: Update freight company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeditionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Spedition"
  /speditionen/{id}/permissions:
    get:
      operationId: listSpeditionPermissions
      tags: [speditionen]
      summary: Get permissions granted by this spedition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SpeditionPermission"
    post:
      operationId: setSpeditionPermission
      tags: [speditionen]
      summary: Set permission for another spedition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpeditionPermissionInput"
      responses:
        "200":
          description: Permission set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeditionPermission"
  /speditionen/{id}/permissions/{receivingId}:
    delete:
      operationId: revokeSpeditionPermission
      tags: [speditionen]
      summary: Revoke permission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: receivingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Revoked
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, email, role, isActive, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [comet_admin, comet_leitstand, comet_lager, comet_viewer, speditions_admin, speditions_bearbeiter, speditions_viewer]
        speditionId:
          type: ["integer", "null"]
        speditionName:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    UserInput:
      type: object
      required: [username, email, password, role]
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [comet_admin, comet_leitstand, comet_lager, comet_viewer, speditions_admin, speditions_bearbeiter, speditions_viewer]
        speditionId:
          type: ["integer", "null"]
        isActive:
          type: boolean
    UserUpdate:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: ["string", "null"]
        role:
          type: string
          enum: [comet_admin, comet_leitstand, comet_lager, comet_viewer, speditions_admin, speditions_bearbeiter, speditions_viewer]
        speditionId:
          type: ["integer", "null"]
        isActive:
          type: boolean
    Spedition:
      type: object
      required: [id, name, kuerzel, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        kuerzel:
          type: string
        ansprechpartner:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        telefon:
          type: ["string", "null"]
        status:
          type: string
          enum: [aktiv, inaktiv]
        bemerkungen:
          type: ["string", "null"]
        createdAt:
          type: string
    SpeditionInput:
      type: object
      required: [name, kuerzel]
      properties:
        name:
          type: string
        kuerzel:
          type: string
        ansprechpartner:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        telefon:
          type: ["string", "null"]
        status:
          type: string
          enum: [aktiv, inaktiv]
        bemerkungen:
          type: ["string", "null"]
    SpeditionUpdate:
      type: object
      properties:
        name:
          type: string
        kuerzel:
          type: string
        ansprechpartner:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        telefon:
          type: ["string", "null"]
        status:
          type: string
          enum: [aktiv, inaktiv]
        bemerkungen:
          type: ["string", "null"]
    SpeditionPermission:
      type: object
      required: [grantingSpeditionId, receivingSpeditionId, permissionLevel]
      properties:
        grantingSpeditionId:
          type: integer
        receivingSpeditionId:
          type: integer
        receivingSpeditionName:
          type: ["string", "null"]
        permissionLevel:
          type: string
          enum: [view, edit]
    SpeditionPermissionInput:
      type: object
      required: [receivingSpeditionId, permissionLevel]
      properties:
        receivingSpeditionId:
          type: integer
        permissionLevel:
          type: string
          enum: [view, edit]
...