# lib/api-zod/src/generated/types/speditionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * COMET LKW-Verladungsverwaltung API
 * OpenAPI spec version: 0.1.0
 */
import type { SpeditionInputStatus } from './speditionInputStatus';

export interface SpeditionInput {
  name: string;
  kuerzel: string;
  /** @nullable */
  ansprechpartner?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  telefon?: string | null;
  status?: SpeditionInputStatus;
  /** @nullable */
  bemerkungen?: string | null;
}